Moving to NPM with Browserify, for single file-bundle
