+++ .eslintrc.json
@@ -1,0 +1,9 @@
+{
+    "parserOptions": {
+      "ecmaVersion": 2020
+    },
+
+    "env": {
+        "es6": true
+    }
+}
