+++ examples/ring/ring.yaml
