+++ rollup.config.js
@@ -1,0 +1,11 @@
+export default [
+  {
+    inlineDynamicImports: true,
+    input: 'src/floorplan-card.js',
+    output: {
+      dir: 'dist',
+      format: 'esm',
+      sourcemap: true
+    },
+  },
+];
